replace unsigned long with size_t

+++ lib_nn/src/c/slice_memcpy.c
@@ -2,7 +2,7 @@
 
 void slice_memcpy(int8_t *dst, int8_t *src, int32_t *in_offsets,
                   int32_t *out_offsets, int32_t *begin, int32_t *end,
-                  void (*memcpy_func)(void *, const void *, unsigned long)) {
+                  void (*memcpy_func)(void *, const void *, size_t)) {
   for (int i0 = begin[0]; i0 < end[0]; i0++) {
     const int32_t in_idx0 = i0 * in_offsets[0];
     const int32_t out_idx0 = (i0 - begin[0]) * out_offsets[0];
